Created PasswordInput component
