+++ layout/ResetPassword/components/ResetPasswordForm/index.tsx
@@ -9,10 +9,7 @@ import Button from "@components/Button";
 import ImageButton from "@components/ImageButton";
 
 import Form from "@components/Form";
-import Input from "@components/Input";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faEye, faEyeSlash } from "@fortawesome/free-solid-svg-icons";
+import PasswordInput from "@components/PasswordInput";
 
 export default function ResetPasswordForm() {
   const router = useRouter();
@@ -23,9 +20,6 @@ export default function ResetPasswordForm() {
   const passwordRef = useRef(null);
   const passwordConfirmationRef = useRef(null);
   const [errorMsg, updateErrorMsg] = useState(null);
-  const [isPasswordVisible, setIsPasswordVisible] = useState(false);
-  const [isConfirmPasswordVisible, setIsConfirmPasswordVisible] =
-    useState(false);
 
   function onSubmit(event) {
     event.preventDefault();
@@ -53,50 +47,22 @@ export default function ResetPasswordForm() {
       });
   }
 
-  const togglePasswordVisibility = () => {
-    setIsPasswordVisible(!isPasswordVisible);
-  };
-
-  const toggleConfirmPasswordVisibility = () => {
-    setIsConfirmPasswordVisible(!isConfirmPasswordVisible);
-  };
-
   return (
     <>
       {success === null && (
         <Form onSubmit={onSubmit}>
-          <Input
+          <PasswordInput
             text="PASSWORD"
-            id="password"
-            name="password"
-            type={isPasswordVisible ? "text" : "password"}
-            autoComplete="current-password"
             fgColor="white"
             bgColor="primary"
-            right={
-              <FontAwesomeIcon
-                className="mx-2"
-                onClick={togglePasswordVisibility}
-                icon={isPasswordVisible ? faEyeSlash : faEye}
-              />
-            }
             ref={passwordRef}
           />
-          <Input
+          <PasswordInput
             text="CONFIRM PASSWORD"
             id="confirm"
             name="confirm"
-            type={isConfirmPasswordVisible ? "text" : "password"}
-            autoComplete="current-password"
             fgColor="white"
             bgColor="primary"
-            right={
-              <FontAwesomeIcon
-                className="mx-2"
-                onClick={toggleConfirmPasswordVisibility}
-                icon={isConfirmPasswordVisible ? faEyeSlash : faEye}
-              />
-            }
             ref={passwordConfirmationRef}
           />
           <p className="mt-10 mb-10 text-center font-iregular text-failure">
