+++ layout/SignUp/components/SignUpForm/index.tsx
@@ -5,11 +5,9 @@ import { useAuth } from "@context/Auth";
 import Button from "@components/Button";
 import Form from "@components/Form";
 import Input from "@components/Input";
+import PasswordInput from "@components/PasswordInput";
 
 import BarebonesQRScanner from "@components/QRScanner/BarebonesQRScanner";
-
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
-import { faEye, faEyeSlash } from "@fortawesome/free-solid-svg-icons";
 
 export default function SignUpForm() {
   const { sign_up, isLoading, errors } = useAuth();
@@ -20,10 +18,6 @@ export default function SignUpForm() {
   const [password, updatePassword] = useState("");
   const [password_confirmation, updatePasswordConfirmation] = useState("");
   const [uuid, setUUID] = useState();
-
-  const [isPasswordVisible, setIsPasswordVisible] = useState(false);
-  const [isConfirmPasswordVisible, setIsConfirmPasswordVisible] =
-    useState(false);
 
   const [local_error, updateError] = useState("");
   const [scanned, updateScanned] = useState(false);
@@ -70,14 +64,6 @@ export default function SignUpForm() {
     }
   };
 
-  const togglePasswordVisibility = () => {
-    setIsPasswordVisible(!isPasswordVisible);
-  };
-
-  const toggleConfirmPasswordVisibility = () => {
-    setIsConfirmPasswordVisible(!isConfirmPasswordVisible);
-  };
-
   return (
     <>
       <Form onSubmit={onFinish}>
@@ -107,39 +93,15 @@ export default function SignUpForm() {
           bgColor="primary"
           onChange={(e) => updateNickname(e.currentTarget.value)}
         />
-        <Input
+        <PasswordInput
           text="PASSWORD"
-          id="password"
-          name="password"
-          type={isPasswordVisible ? "text" : "password"}
-          autoComplete="current-password"
           fgColor="white"
           bgColor="primary"
-          onChange={(e) => updatePassword(e.currentTarget.value)}
-          right={
-            <FontAwesomeIcon
-              className="mx-2"
-              onClick={togglePasswordVisibility}
-              icon={isPasswordVisible ? faEyeSlash : faEye}
-            />
-          }
         />
-        <Input
+        <PasswordInput
           text="CONFIRM PASSWORD"
-          id="password"
-          name="password"
-          type={isConfirmPasswordVisible ? "text" : "password"}
-          autoComplete="current-password"
           fgColor="white"
           bgColor="primary"
-          onChange={(e) => updatePasswordConfirmation(e.currentTarget.value)}
-          right={
-            <FontAwesomeIcon
-              className="mx-2"
-              onClick={toggleConfirmPasswordVisibility}
-              icon={isConfirmPasswordVisible ? faEyeSlash : faEye}
-            />
-          }
         />
         <Button
           text={scanning ? "STOP SCANNING" : "SCAN QR"}
